[chore]: update folder test
